+++ packages/tizen_app_manager/lib/tizen_app_manager.dart
@@ -2,4 +2,4 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-export 'app_manager.dart';
+export 'src/app_manager.dart';